Added class SelectOption
